+++ ceammc/ext/src/system/mod_system.cpp
@@ -7,6 +7,7 @@ void setup_system_exec();
 void setup_system_exit();
 void setup_system_getenv();
 void setup_system_hostname();
+void setup_system_info();
 void setup_system_memsize();
 void setup_system_memused();
 void setup_system_screen_size();
@@ -20,6 +21,7 @@ void ceammc_system_setup()
     setup_system_exit();
     setup_system_getenv();
     setup_system_hostname();
+    setup_system_info();
     setup_system_memsize();
     setup_system_memused();
     setup_system_screen_size();